requirements.txt
python-dotenv
langchain-google-genai
langchain-core
PyPDF2
python-pptx
pyfiglet
pytesseract
pdf2image
pillow
google-generativeai
